migrate rule
